third_party/blink/web_tests/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-seed-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'seed' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.seed.baseVal is 10
PASS turbulence.seed.baseVal is 5
PASS successfullyParsed is true

TEST COMPLETE
